+++ projects/angular-cesium/src/lib/angular-cesium-widgets/services/entity-editors/ellipses-editor/ellipses-editor.service.ts
@@ -41,7 +41,7 @@ export const DEFAULT_ELLIPSE_OPTIONS: EllipseEditOptions = {
     shadows: Cesium.ShadowMode.DISABLED,
   },
   pointProps: {
-    color: Cesium.Color.WHITE.withAlpha(0.95),
+    color: Cesium.Color.WHITE,
     outlineColor: Cesium.Color.BLACK.withAlpha(0.2),
     outlineWidth: 1,
     pixelSize: 13,