+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+### Features
+* Add more stackblitz examples and refactor the demo examples.
+* Change Circle, Ellipse and Hippodrome default point color changed to white. 
+
 ## 0.61
 ### Features
 * Add `PickConfiguration` for `MapEventManager` for setting `pick()` width, height and `drillPick()` limit.